# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TwinBlueprint Server API",
    "description": "REST API for TwinBlueprint — leads, bids, projects, campaigns, analytics, and admin auth.",
    "version": "3.0.0",
    "contact": {
      "name": "TwinBlueprint",
      "email": "twinblueprints@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token. Can also be sent via cookie named 'token'."
      }
    },
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "project": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "start_date": {
            "type": "string",
            "description": "Date string (YYYY-MM-DD)"
          },
          "end_date": {
            "type": "string",
            "description": "Date string (YYYY-MM-DD)"
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "suppliers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "uses_3d": {
            "type": "boolean"
          },
          "competitor": {
            "type": "string",
            "nullable": true
          },
          "issue": {
            "type": "string",
            "nullable": true,
            "description": "Single issues text field, not an array."
          },
          "bid_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "In-flight",
              "On Hold",
              "Completed",
              "Cancelled"
            ]
          },
          "phase": {
            "type": "string",
            "enum": [
              "Planning",
              "Design",
              "Construction",
              "In Progress",
              "Completed"
            ]
          },
          "value": {
            "type": "number",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "project",
          "client",
          "start_date",
          "end_date",
          "progress",
          "suppliers",
          "uses_3d",
          "created_at"
        ]
      },
      "CreateProjectInput": {
        "type": "object",
        "properties": {
          "project": {
            "type": "string",
            "example": "Miami Worldcenter"
          },
          "client": {
            "type": "string",
            "example": "MDM Group"
          },
          "start_date": {
            "type": "string",
            "example": "2024-03-01"
          },
          "end_date": {
            "type": "string",
            "example": "2027-08-01"
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "suppliers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "uses_3d": {
            "type": "boolean"
          },
          "competitor": {
            "type": "string",
            "nullable": true
          },
          "issue": {
            "type": "string",
            "nullable": true,
            "description": "Single issues text field, not an array."
          },
          "bid_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "In-flight",
              "On Hold",
              "Completed",
              "Cancelled"
            ],
            "default": "Active"
          },
          "phase": {
            "type": "string",
            "enum": [
              "Planning",
              "Design",
              "Construction",
              "In Progress",
              "Completed"
            ],
            "default": "Planning"
          },
          "value": {
            "type": "number",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "project",
          "client",
          "start_date",
          "end_date"
        ]
      },
      "UpdateProjectInput": {
        "type": "object",
        "properties": {
          "project": {
            "type": "string"
          },
          "client": {
            "type": "string"
          },
          "start_date": {
            "type": "string"
          },
          "end_date": {
            "type": "string"
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "suppliers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "uses_3d": {
            "type": "boolean"
          },
          "competitor": {
            "type": "string",
            "nullable": true
          },
          "issue": {
            "type": "string",
            "nullable": true,
            "description": "Single issues text field, not an array."
          },
          "bid_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "Active",
              "In-flight",
              "On Hold",
              "Completed",
              "Cancelled"
            ]
          },
          "phase": {
            "type": "string",
            "enum": [
              "Planning",
              "Design",
              "Construction",
              "In Progress",
              "Completed"
            ]
          },
          "value": {
            "type": "number",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "pages": {
            "type": "integer"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "string",
            "description": "Only included in development"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "region": "Region is required"
            }
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/projects": {
      "post": {
        "tags": [
          "Projects"
        ],
        "summary": "Create a project",
        "description": "Admin only. Requires project, client, start_date, end_date. Saves progress (0–100), supplier names, uses_3d, competitor, issue, bid_id and status. Status defaults to Active; only In-flight appears in pipeline inflight_projects.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "project": {
                          "$ref": "#/components/schemas/Project"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "List all projects",
        "description": "Supports pagination.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "projects": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Project"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get project by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "project": {
                          "$ref": "#/components/schemas/Project"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Update a project",
        "description": "Admin only. Send one or more project fields. Omitted fields are preserved. Allowed status values: Active, In-flight, On Hold, Completed, Cancelled. Refetch GET /api/pipeline after saving; only In-flight projects appear in inflight_projects, with all saved fields.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "project": {
                          "$ref": "#/components/schemas/Project"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete a project",
        "description": "Admin only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{id}/suppliers/{supplierId}": {
      "put": {
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Link supplier to project (admin, idempotent)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link saved"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Project or supplier not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Unlink supplier from project (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "supplierId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Link saved"
          },
          "400": {
            "description": "Invalid UUID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin only"
          },
          "404": {
            "description": "Project or supplier not found"
          }
        }
      }
    },
...
  }
}